adapt SMS strings in preferences
Fixes #2672
Closes #2682

//FREEBIE
