+++ res/xml/preferences.xml
@@ -1,12 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <CheckBoxPreference android:defaultValue="false"
-                        android:key="pref_toggle_push_messaging"
-                        android:title="@string/preferences__use_data_channel"
-                        android:summary="@string/preferences__use_the_data_channel_for_communication_with_other_textsecure_users"
-                        android:icon="?pref_ic_push"/>
-
     <Preference android:key="preference_category_sms_mms"
                 android:title="@string/preferences__sms_mms"
                 android:icon="?pref_ic_sms_mms"/>
