+++ res/xml/preferences_advanced.xml
@@ -2,6 +2,11 @@
 
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
     <CheckBoxPreference android:defaultValue="false"
+                        android:key="pref_toggle_push_messaging"
+                        android:title="@string/preferences__textsecure_messages"
+                        android:summary="@string/preferences__use_the_data_channel_for_communication_with_other_textsecure_users"/>
+
+    <CheckBoxPreference android:defaultValue="false"
                         android:key="pref_enter_key"
                         android:title="@string/preferences__enable_enter_key_title"
                         android:summary="@string/preferences__replace_smiley_with_enter_key"/>
