+++ res/xml/preferences_sms_mms.xml
@@ -12,8 +12,8 @@
                         android:title="@string/preferences__pref_all_mms_title" />
 
     <Preference android:key="pref_set_default"
-                android:title="@string/preferences__make_default_sms_app"
-                android:summary="@string/preferences__make_textsecure_the_default_sms_mms_app" />
+                android:title="@string/ApplicationPreferencesActivity_sms_disabled"
+                android:summary="@string/ApplicationPreferencesActivity_touch_to_make_textsecure_your_default_sms_app" />
 
     <CheckBoxPreference android:defaultValue="false"
                         android:key="pref_delivery_report_sms"
