+++ src/org/thoughtcrime/securesms/preferences/AdvancedPreferenceFragment.java
@@ -1,24 +1,41 @@
 package org.thoughtcrime.securesms.preferences;
 
 import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.preference.CheckBoxPreference;
 import android.preference.Preference;
 import android.provider.ContactsContract;
 import android.support.v4.preference.PreferenceFragment;
 import android.util.Log;
+import android.widget.Toast;
+
+import com.google.android.gms.gcm.GoogleCloudMessaging;
 
 import org.thoughtcrime.securesms.ApplicationPreferencesActivity;
 import org.thoughtcrime.securesms.LogSubmitActivity;
 import org.thoughtcrime.securesms.R;
+import org.thoughtcrime.securesms.RegistrationActivity;
 import org.thoughtcrime.securesms.contacts.ContactAccessor;
 import org.thoughtcrime.securesms.contacts.ContactIdentityManager;
+import org.thoughtcrime.securesms.push.TextSecureCommunicationFactory;
+import org.thoughtcrime.securesms.util.ProgressDialogAsyncTask;
+import org.thoughtcrime.securesms.util.ResUtil;
 import org.thoughtcrime.securesms.util.TextSecurePreferences;
+import org.whispersystems.libaxolotl.util.guava.Optional;
+import org.whispersystems.textsecure.api.TextSecureAccountManager;
+import org.whispersystems.textsecure.api.push.exceptions.AuthorizationFailedException;
+
+import java.io.IOException;
 
 public class AdvancedPreferenceFragment extends PreferenceFragment {
   private static final String TAG = AdvancedPreferenceFragment.class.getSimpleName();
 
+  private static final String PUSH_MESSAGING_PREF   = "pref_toggle_push_messaging";
   private static final String SUBMIT_DEBUG_LOG_PREF = "pref_submit_debug_logs";
 
   private static final int PICK_IDENTITY_CONTACT = 1;
@@ -28,6 +45,7 @@ public class AdvancedPreferenceFragment extends PreferenceFragment {
     super.onCreate(paramBundle);
     addPreferencesFromResource(R.xml.preferences_advanced);
 
+    initializePushMessagingToggle();
     initializeIdentitySelection();
 
     this.findPreference(SUBMIT_DEBUG_LOG_PREF)
@@ -48,6 +66,12 @@ public class AdvancedPreferenceFragment extends PreferenceFragment {
     if (resultCode == Activity.RESULT_OK && reqCode == PICK_IDENTITY_CONTACT) {
       handleIdentitySelection(data);
     }
+  }
+
+  private void initializePushMessagingToggle() {
+    CheckBoxPreference preference = (CheckBoxPreference)this.findPreference(PUSH_MESSAGING_PREF);
+    preference.setChecked(TextSecurePreferences.isPushRegistered(getActivity()));
+    preference.setOnPreferenceChangeListener(new PushMessagingClickListener());
   }
 
   private void initializeIdentitySelection() {
@@ -95,6 +119,84 @@ public class AdvancedPreferenceFragment extends PreferenceFragment {
       final Intent intent = new Intent(getActivity(), LogSubmitActivity.class);
       startActivity(intent);
       return true;
+    }
+  }
+
+  private class PushMessagingClickListener implements Preference.OnPreferenceChangeListener {
+    private static final int SUCCESS       = 0;
+    private static final int NETWORK_ERROR = 1;
+
+    private class DisablePushMessagesTask extends ProgressDialogAsyncTask<Void, Void, Integer> {
+      private final CheckBoxPreference checkBoxPreference;
+
+      public DisablePushMessagesTask(final CheckBoxPreference checkBoxPreference) {
+        super(getActivity(), R.string.ApplicationPreferencesActivity_unregistering, R.string.ApplicationPreferencesActivity_unregistering_for_data_based_communication);
+        this.checkBoxPreference = checkBoxPreference;
+      }
+
+      @Override
+      protected void onPostExecute(Integer result) {
+        super.onPostExecute(result);
+        switch (result) {
+        case NETWORK_ERROR:
+          Toast.makeText(getActivity(),
+                         R.string.ApplicationPreferencesActivity_error_connecting_to_server,
+                         Toast.LENGTH_LONG).show();
+          break;
+        case SUCCESS:
+          checkBoxPreference.setChecked(false);
+          TextSecurePreferences.setPushRegistered(getActivity(), false);
+          break;
+        }
+      }
+
+      @Override
+      protected Integer doInBackground(Void... params) {
+        try {
+          Context                  context        = getActivity();
+          TextSecureAccountManager accountManager = TextSecureCommunicationFactory.createManager(context);
+
+          accountManager.setGcmId(Optional.<String>absent());
+          GoogleCloudMessaging.getInstance(context).unregister();
+
+          return SUCCESS;
+        } catch (AuthorizationFailedException afe) {
+          Log.w(TAG, afe);
+          return SUCCESS;
+        } catch (IOException ioe) {
+          Log.w(TAG, ioe);
+          return NETWORK_ERROR;
+        }
+      }
+    }
+
+    @Override
+    public boolean onPreferenceChange(final Preference preference, Object newValue) {
+      if (((CheckBoxPreference)preference).isChecked()) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        builder.setIcon(ResUtil.getDrawable(getActivity(), R.attr.dialog_info_icon));
+        builder.setTitle(R.string.ApplicationPreferencesActivity_disable_push_messages);
+        builder.setMessage(R.string.ApplicationPreferencesActivity_this_will_disable_push_messages);
+        builder.setNegativeButton(android.R.string.cancel, null);
+        builder.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
+          @Override
+          public void onClick(DialogInterface dialog, int which) {
+            new DisablePushMessagesTask((CheckBoxPreference)preference).execute();
+          }
+        });
+        builder.show();
+      } else {
+        Intent nextIntent = new Intent(getActivity(), ApplicationPreferencesActivity.class);
+        nextIntent.putExtra("master_secret", getActivity().getIntent().getParcelableExtra("master_secret"));
+
+        Intent intent = new Intent(getActivity(), RegistrationActivity.class);
+        intent.putExtra("cancel_button", true);
+        intent.putExtra("next_intent", nextIntent);
+        intent.putExtra("master_secret", getActivity().getIntent().getParcelableExtra("master_secret"));
+        startActivity(intent);
+      }
+
+      return false;
     }
   }
 }
